requirements.txt
numpy==1.23.5
Pillow==9.5.0
requests==2.25.1
pymongo==3.11.3
Flask==2.3.0
pytest==7.3.2
mysql-connector-python==8.1.0
flask-restx==1.1.0
flake8==6.1.0
black==23.7.0
